Better (?) patchcord editing
Still a bit weird in places... there's debug code in at the moment, too

git-svn-id: svn://slug1/jonathan/SVNrepo/AudioPatcher/trunk@18 f6964fe3-6982-4f03-a74d-735ce54faba2

+++ objects.h
@@ -150,7 +150,7 @@ class AudioObjInstance
     AudioObjPtr_u streamP;
     int16_t x;
     int16_t y;
-    uint16_t inputAvailFlags;
+    uint32_t inputAvailFlags; // let's be optimistic!
 };
 
 struct AudioObjInstancePtr {AudioObjInstance* p; };
